refactor: split functional code
